FROMLIST: drm/vkms: Support multiple DRM objects (crtcs, etc.) per VKMS device
This change supports multiple CRTCs, encoders, connectors instead of one
of each per device.

Since ConfigFS-based devices will support multiple crtcs, it's useful to
move all of the writeback/composition data from being per-"output" to
being per-CRTC.

Since there's still only ever one CRTC, this should be a no-op refactor.

Signed-off-by: Jim Shargo <jshargo@chromium.org>
Signed-off-by: Brandon Pollack <brpol@chromium.org>
(am from https://patchwork.kernel.org/project/dri-devel/patch/20230829053201.423261-3-brpol@chromium.org/)

The change is in the final stages of review (fifth iteration), All that
remains is some (potential) igt tests and final signoff.

Some reasons we'd like to submit in chromeOS are:

This will allow our virtual multi display tests to run in presubmit.
This real world case will be compelling evidence if/when upstream
requests some data that this ran for real in production (and avoiding
the chicken egg problem of we need this in the kernel to run the tests
but without running the tests we can't submit upstream).

Change-Id: Id2325f1ee3bd7fc588c87443e7bda6481416a1cd
Signed-off-by: Brandon Pollack <brpol@chromium.org>
BUG=241800059
BUG=283357160
TEST=new tests defined here passing https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4666669
Signed-off-by: Brandon Pollack <brpol@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4810099
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>