FROMLIST: drm/vkms: Add a module param to enable/disable the default device
In many testing circumstances, we will want to just create a new device
and test against that. If we create a default device, it can be annoying
to have to manually select the new device instead of choosing the only
one that exists.

The param, enable_default, is defaulted to true to maintain backwards
compatibility.

Signed-off-by: Jim Shargo <jshargo@chromium.org>
Signed-off-by: Brandon Pollack <brpol@chromium.org>
(am from https://patchwork.kernel.org/project/dri-devel/patch/20230829053201.423261-7-brpol@chromium.org/)

The change is in the final stages of review (fifth iteration), All that
remains is some (potential) igt tests and final signoff.

Some reasons we'd like to submit in chromeOS are:

This will allow our virtual multi display tests to run in presubmit.
This real world case will be compelling evidence if/when upstream
requests some data that this ran for real in production (and avoiding
the chicken egg problem of we need this in the kernel to run the tests
but without running the tests we can't submit upstream).

Signed-off-by: Brandon Pollack <brpol@chromium.org>
BUG=241800059
TEST=new tests defined here passing https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/4666669
BUG=283357160
Change-Id: Ib9605709186c45b526aa5a81b5b5686f62b5f55a
Signed-off-by: Brandon Pollack <brpol@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4810103
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>