UPSTREAM: ACPI: EC: Fix acpi_ec_dispatch_gpe()
Commit 896e97bf99ec ("ACPI: EC: Clear GPE on interrupt handling only")
broke suspend-to-idle at least on Dell XPS13 9360 and 9380.

The problem is that acpi_ec_dispatch_gpe() must clear the EC GPE,
because the EC GPE handler never runs when the system is in the
suspend-to-idle state and if the EC GPE is not cleared by the suspend-
to-idle loop, it is never cleared at all which leads to a GPE storm.
This causes suspend-to-idle to burn energy instead of saving it which
is potentially dangerous (the affected machines heat up rather badly
when that happens).

Addess this by making acpi_ec_dispatch_gpe() clear the EC GPE as it did
before.

Fixes: 896e97bf99ec ("ACPI: EC: Clear GPE on interrupt handling only")
Tested-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Change-Id: Ib61bf2fdebdcdd80f29bdc946cdf7644b9edf9db
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Link: https://lore.kernel.org/lkml/12234124.O9o76ZdvQC@kreacher/
(cherry-picked from commit b5539eb5ee70257520e40bb636a295217c329a50 https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git)
TEST=/sys/firmware/acpi/interrupts/sci_not stays at 0 on rex
BUG=b:299648344
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4854049
Reviewed-by: Paz Zcharya <pazz@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Anil Kumar K <anil.kumar.k@intel.corp-partner.google.com>