UPSTREAM: drm/i915/gt: Clear all bits from GEN12_FF_MODE2
Right now context workarounds don't do a rmw and instead only write to
the register. Since 2 separate programmings to the same register are
coalesced into a single write, this is not problematic for
GEN12_FF_MODE2 since both TDS and GS timer are going to be written
together and the other remaining bits be zeroed.

However in order to fix other workarounds that may want to preserve the
unrelated bits in the same register, context workarounds need to
be changed to a rmw. To prepare for that, move the programming of
GEN12_FF_MODE2 to a single place so the value passed for "clear" can
be all the bits. Otherwise the second workaround would be dropped as
it'd be detected as overwriting a previously programmed workaround.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230630203509.1635216-3-lucas.demarchi@intel.com
(cherry picked from commit e8f7df163cb40152b757be0f66a9c6b520b3037a)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:293823364
TEST= No intel_gt_init error while running warm reboot stress.

Change-Id: I3f0fc73265e163cf41b858d1e30f413d293e234a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4792826
Tested-by: Kun Liu <liukun11@huaqin.corp-partner.google.com>
Commit-Queue: Shawn C Lee <shawn.c.lee@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>