CHROMIUM: media: camx: properly put mapped buffer
First of all, we need to put the buffer, otherwise we leak
its ref-count.

However, depending on the CAM_CDM_BL_CMD_TYPE we may have a
mapped mem-handle buffer (which we need to put) or iova, which
we cannot put. Therefore we need to test CAM_CDM_BL_CMD_TYPE
when we don't need the buffer anymore. Factor two helper
function that take of that.

Besides we need to put the buffer on error-out paths; this
is also fixed.

UPSTREAM-TASK=b:274966069
BUG=b:287178085
TEST=CTS on quackingstick

Change-Id: Iba8abcb7bf83b33d7f806da32b0bf7e3ec429e94
Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4686558
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>