UPSTREAM: drm/msm: Fix hw_fence error path cleanup
In an error path where the submit is free'd without the job being run,
the hw_fence pointer is simply a kzalloc'd block of memory.  In this
case we should just kfree() it, rather than trying to decrement it's
reference count.  Fortunately we can tell that this is the case by
checking for a zero refcount, since if the job was run, the submit would
be holding a reference to the hw_fence.

Fixes: f94e6a51e17c ("drm/msm: Pre-allocate hw_fence")
Signed-off-by: Rob Clark <robdclark@chromium.org>
Patchwork: https://patchwork.freedesktop.org/patch/547088/
(cherry picked from commit 1cd0787f082e1a179f2b6e749d08daff1a9f6b1b)

BUG=b:300094055
TEST=build test

Change-Id: Ibd0432044e24729758ee08553bb3d1d80e2d8ce5
Signed-off-by: Drew Davenport <ddavenport@chromium.org>